Split route, response handlers into proper modules
