+++ lib/otto/route_handlers/factory.rb
@@ -1,0 +1,29 @@
+# frozen_string_literal: true
+
+# lib/otto/route_handlers/factory.rb
+
+require_relative 'base'
+
+class Otto
+  module RouteHandlers
+    # Factory for creating appropriate handlers based on route definitions
+    class HandlerFactory
+      # Create a handler for the given route definition
+      # @param route_definition [Otto::RouteDefinition] The route definition
+      # @param otto_instance [Otto] The Otto instance for configuration access
+      # @return [BaseHandler] Appropriate handler for the route
+      def self.create_handler(route_definition, otto_instance = nil)
+        case route_definition.kind
+        when :logic
+          LogicClassHandler.new(route_definition, otto_instance)
+        when :instance
+          InstanceMethodHandler.new(route_definition, otto_instance)
+        when :class
+          ClassMethodHandler.new(route_definition, otto_instance)
+        else
+          raise ArgumentError, "Unknown handler kind: #{route_definition.kind}"
+        end
+      end
+    end
+  end
+end
